+++ core/datasources-service/lib/utils/Repository.js
@@ -7,7 +7,6 @@ const normalize = require('./normalize');
 const dvcConfig = require('./dvcConfig');
 const getFilePath = require('./getFilePath');
 const DvcClient = require('./DvcClient');
-const { timeWrap } = require('./timeWarp');
 
 /**
  * @typedef {import('./types').FileMeta} FileMeta
@@ -49,8 +48,6 @@ class Repository {
         this.rootDir = rootDir;
         this.dvc = new DvcClient(this.cwd, this.repositoryUrl);
         this.generateDvcConfig = dvcConfig(this.config);
-        ['_setupDvcRepository', '_enrichDvcFile', 'setup', 'ensureClone', 'addFiles', 'pullFiles', 'moveExistingFiles', 'scanDir', 'dropNonDataFiles', 'dropFiles', 'loadMetaDataFiles', 'push', 'cleanup', 'filterFilesFromClone', 'filterMetaFilesFromClone', 'deleteClone'].forEach(m => this[m] = timeWrap(this[m], this));
-
     }
 
     /** @param {string} name */
